Improve performance by caching metadata fields
This assumes that the fields won't be changed except when this
program tries to change them, but refreshes all fields when it
changes any field. If you externally change a value, it isn't
shown on the gui unless you update any value on the gui, which
can be the externally-changed value, losing external edits.
But refreshing the entire repository reloads everything including
any external changes that weren't overwritten like this.